# public/openapi.json
{"openapi":"3.0.0","paths":{"/ping":{"get":{"operationId":"AppController_ping","summary":"Return OK","parameters":[],"responses":{"default":{"description":"","content":{"application/json":{"schema":{"type":"string"}}}}}}},"/records":{"get":{"operationId":"RecordController_find","summary":"Return a list of records by opt","parameters":[{"name":"select","required":false,"in":"query","schema":{"type":"array","items":{"type":"string"}}},{"name":"where","required":false,"in":"query","schema":{"type":"object"}},{"name":"order","required":false,"in":"query","schema":{"type":"object"}},{"name":"skip","required":false,"in":"query","schema":{"type":"number"}},{"name":"take","required":false,"in":"query","schema":{"type":"number"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Record"}}}}}},"security":[{"authorization":[]}]},"patch":{"operationId":"RecordController_update","summary":"Update a record by where","parameters":[{"name":"id","required":false,"in":"query","schema":{"type":"number"}},{"name":"name","required":false,"in":"query","schema":{"type":"string"}},{"name":"content","required":false,"in":"query","schema":{"type":"string"}},{"name":"location","required":false,"in":"query","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RecordDto"}}}},"responses":{"204":{"description":""}},"security":[{"authorization":[]}]},"post":{"operationId":"RecordController_create","summary":"Create a record","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RecordDto"}}}},"responses":{"201":{"description":""}},"security":[{"authorization":[]}]}},"/records/count":{"get":{"operationId":"RecordController_count","summary":"Return a count number of records by where","parameters":[{"name":"id","required":false,"in":"query","schema":{"type":"number"}},{"name":"name","required":false,"in":"query","schema":{"type":"string"}},{"name":"content","required":false,"in":"query","schema":{"type":"string"}},{"name":"location","required":false,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"number"}}}}},"security":[{"authorization":[]}]}},"/records/{id}":{"get":{"operationId":"RecordController_findById","summary":"Return a record by id","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Record"}}}}},"security":[{"authorization":[]}]},"patch":{"operationId":"RecordController_updateById","summary":"Update a record by id","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RecordDto"}}}},"responses":{"204":{"description":""}},"security":[{"authorization":[]}]},"delete":{"operationId":"RecordController_deleteById","summary":"Delete a record by id","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"204":{"description":""}},"security":[{"authorization":[]}]}},"/records/upsert":{"post":{"operationId":"RecordController_upsert","summary":"Upsert records","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"array","items":{"type":"string"}}}}},"responses":{"204":{"description":""}},"security":[{"authorization":[]}]}},"/records/getByLike":{"post":{"operationId":"RecordController_getByLike","summary":"Return a list of records by like","parameters":[{"name":"content","required":false,"in":"query","schema":{"type":"string"}},{"name":"name","required":false,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Record"}}}}}},"security":[{"authorization":[]}]}},"/records/deleteAndCreate/{id}":{"post":{"operationId":"RecordController_deleteAndCreate","summary":"Delete a record and create a record","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RecordDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"string"}}}}},"security":[{"authorization":[]}]}}},"info":{"title":"nest-api","description":"Generated with `@nest/swagger`","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"authorization":{"scheme":"bearer","bearerFormat":"JWT","type":"apiKey","in":"header","name":"Authorization"}},"schemas":{"FindManyOptionsDto":{"type":"object","properties":{"select":{"type":"array","items":{"type":"string"}},"where":{"type":"object"},"order":{"type":"object"},"skip":{"type":"number"},"take":{"type":"number"}}},"FindOneOptionsDto":{"type":"object","properties":{"select":{"type":"array","items":{"type":"string"}},"where":{"type":"object"},"order":{"type":"object"}}},"Record":{"type":"object","properties":{"id":{"type":"number"},"name":{"type":"string"},"content":{"type":"string"},"location":{"type":"string"},"createTime":{"type":"string"},"updateTime":{"type":"string"}},"required":["id","name","content","createTime","updateTime"]},"RecordDto":{"type":"object","properties":{"id":{"type":"number"},"name":{"type":"string"},"content":{"type":"string"},"location":{"type":"string"}}}}}}